Fix: 생성자 타입 변환 적용 안되었었음
생성자 타입 변환 미적용 및 Status코드 컨플릭트 수정과정에서의 세미콜론 쉼표로 안 바뀜

Fix: #2

+++ backend/src/main/java/com/ssafy/umzip/domain/review/entity/Review.java
@@ -32,7 +32,7 @@ public class Review extends BaseTimeEntity {
     private Member member;
 
     @Builder
-    public Review(double score, String content, Member member) {
+    public Review(float score, String content, Member member) {
         this.score = score;
         this.content = content;
         this.member = member;
